nav update + HP finition
