+++ front/src/App.css
@@ -5,6 +5,7 @@
   font-family: 'Bebas Neue';
   font-size: 24px;
   letter-spacing: 1px;
+  margin-top: 35px;
 }
 
 nav ul li a {
@@ -12,6 +13,7 @@ nav ul li a {
   text-transform: uppercase;
   display: block;
   padding: 15px 25px;
+  text-decoration: none;
 }
 
 #c {
@@ -67,6 +69,17 @@ canvas {
   filter: drop-shadow(0 0 2em #61dafbaa);
 }
 
+.HPcontainer{
+  margin-top: auto;
+  margin-bottom: auto;
+}
+
+#root{
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -80,15 +93,6 @@ canvas {
   -moz-osx-font-smoothing: grayscale;
 }
 
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-
-a:hover {
-  color: #535bf2;
-}
 
 body {
   margin: 0;
@@ -122,14 +126,54 @@ button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
 }
 
+.cardcontainer{
+  display: flex;
+  justify-content: space-around;
+}
+
+.cards{
+  color: white;
+    padding: 25px;
+    min-height: 525px;
+    width: 25%;
+    background-position: center center;
+    background-size: cover;
+    position: relative;
+    cursor: pointer;
+}
+
+.cardcontainercontent{
+  position: absolute;
+  bottom: 0;
+}
+
+.cardcontainercontent a{
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.card1{
+  background-image: url('./assets/CardsHP1.jpeg');
+}
+
+.card2{
+  background-image: url('./assets/CardsHP2.jpeg');
+}
+
+.card3{
+  background-image: url('./assets/CardsHP3.jpeg');
+}
+
+.cardstitle {
+  text-transform: uppercase;
+  font-family: 'Bebas Neue';
+  font-size: 30px;
+}
+
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
     background-color: #ffffff;
-  }
-  
-  a:hover {
-    color: #747bff;
   }
   
   button {
