+++ front/src/routes/AppRouter.tsx
@@ -1,21 +1,30 @@
-// AppRouter.tsx
 import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom';
 import Home from '../pages/Home';
-import Menu from '../components/menu/Menu'; // Assurez-vous que le chemin d'importation est correct
+import Menu from '../components/menu/Menu';
 import ScenarioSelection from '../pages/Scenarioselection';
 import PlayerSelection from '../pages/PlayerSelection';
 
-const AppRouter: React.FC = () => (
-  <Router>
+// 
+
+const AppContent: React.FC = () => {
+  const location = useLocation(); // Obtenez le chemin actuel
+
+  return (
     <>
-      <Menu />
+      {location.pathname !== "/" && <Menu />}
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/scenario" element={<ScenarioSelection />} />
         <Route path="/player" element={<PlayerSelection />} />
       </Routes>
     </>
+  );
+};
+
+const AppRouter: React.FC = () => (
+  <Router>
+    <AppContent />
   </Router>
 );
